Split __main__ and cli functions into two files
